Started refactoring to have states within phases. This makes the logic easier to follow with entering and leaving states especially.
