+++ Rpg.Mobile.App/Game/MainBattle/Components/BattleGridScene.cs
@@ -1,6 +1,6 @@
 ﻿using Rpg.Mobile.App.Game.Common;
 using Rpg.Mobile.App.Game.MainBattle.Events;
-using Rpg.Mobile.App.Game.MainBattle.States;
+using Rpg.Mobile.App.Game.MainBattle.States.SetupPhase;
 using Rpg.Mobile.App.Game.MainBattle.Systems.Calculators;
 using Rpg.Mobile.App.Game.MainBattle.Systems.Data;
 using Rpg.Mobile.App.Utils;
